+++ interfaces/kits/js/src/mod_fileio/properties/fchmod.cpp
@@ -70,7 +70,7 @@ napi_value Fchmod::Async(napi_env env, napi_callback_info info)
         return nullptr;
     }
 
-    int argc = funcArg.GetArgc();
+    size_t argc = funcArg.GetArgc();
     bool succ = false;
     int fd;
     tie(succ, fd) = NVal(env, funcArg[NARG_POS::FIRST]).ToInt32();
