+++ interfaces/kits/js/src/mod_fileio/properties/fsync.cpp
@@ -83,7 +83,7 @@ napi_value Fsync::Async(napi_env env, napi_callback_info info)
         }
     };
     string procedureName = "FileIOFsync";
-    int argc = funcArg.GetArgc();
+    size_t argc = funcArg.GetArgc();
     NVal thisVar(env, funcArg.GetThisVar());
     if (argc == NARG_CNT::ONE) {
         return NAsyncWorkPromise(env, thisVar).Schedule(procedureName, cbExec, cbComplete).val_;
